feat(projection): migrate projection module into package-local package
