+++ packages/projection/src/IEventSubscription.ts
