+++ packages/projection/src/IProjectionStore.ts
