+++ packages/projection/src/InMemoryProjectionStore.ts
