--- src/projections/index.ts
@@ -1,24 +1,0 @@
-export type { Checkpoint, ProjectionEvent, EventBatch, Cursor } from './types';
-export type { IProjectionStore } from './IProjectionStore';
-export type { IEventSubscription } from './IEventSubscription';
-export type { IProjectionLinkStore } from './IProjectionLinkStore';
-export {
-  createProjection
-} from './createProjection';
-export type {
-  AggregateDefinition,
-  AggregateEventPayloadMap,
-  AggregateEventKeys,
-  AggregateEventPayloadByKey,
-  ProjectionContext,
-  ProjectionHandler,
-  ProjectionHandlers,
-  ProjectionStreamDefinition,
-  JoinStreamDefinition,
-  ProjectionDefinition,
-  ProjectionBuilder
-} from './createProjection';
-export { InMemoryProjectionStore } from './InMemoryProjectionStore';
-export { InMemoryProjectionLinkStore } from './InMemoryProjectionLinkStore';
-export { ProjectionDaemon } from './ProjectionDaemon';
-export type { ProjectionDaemonOptions, BatchStats } from './ProjectionDaemon';
